Normally, the CNF proof manager remembers the base assertion from which rewritten assertions originated.
When dealing with splitting lemmas, however, this causes an error, because the un-rewritten version of the lemma does not get encoded into a CNF literal. Later, when the proof manager asks for a literal, the lemma doesn't exist in the mapping.

The solution: when dealing with splitting lemmas (identifiable by their proof rule, RULE_INVALID), act like it doesn't have an un-rewritten version.