Initial support for proofs in cases where more than one theory is active.
The key issue is being able to tell which theory should be called upon for proving each of the necessary lemmas, which may contain literals owned by other theories. To this end, we keep records of which theory asserted/propagated which lemma.

There are two ways lemmas can make their way into the SAT solver:
1. A theory explicitly called d_out->lemma() and asserted the lemma.
   In this case, the lemma flows from the theory into the theory engine, the prop engine and into the cnf stream.
   The cnf stream then registers the lemma and the theory that generated it in cnf_proof.

2. Lemmas can be asserted when the sat solver requests an explanation following a theory propagation.
   In this case, the theory engine's getExplanation() method now tells us which was the theory responsible
   for the last propagation. The theory_proxy then registers this theory with the cnf_proof, and all subsequent
   lemmas asserted to the sat_proof are associated with this theory.

When the time comes to prove a lemma we just look at the records and have that theory produce the proof.