Added support for lemmas owned by the BUILTIN theory. This lemmas appear to occur (only?) as a result of propagations from the shared term database, and their pursuing explanations.
Since the shared term database is similar to an equality engine, these lemmas will be explained by the UF theory.
Here's there's a twist: so far, UF lemmas always included a disequality as part of the input for the contradiction. In the case of the shared term database, contradictions can be derived also from a disequality between constants. So, adjusted the UF prover to support these cases.

For the actual constant disequality, we will need to call upon the owner theory. At present we just pass this request to the theory proof engine for dispatching, but as this has not been implemented yet it just prints a trust statement.