Enable proofs when CVC4 is built with LFSC checker.
Disable building with LFSC for distcheck travis build. LFSC lives
outside of the CVC4 repository and is not part of the CVC4 distribution
anymore. Further, make distcheck with LFSC would require to call
contrib/get-lfsc-checker before calling make check on the temp build
(the build of the unpacked distribution tar ball).