Add gcc7 jobs to Travis
This commit adds two jobs (debug, with portfolio, test groups 0 and 1) that are
allowed to fail to Travis. The jobs are allowed to fail because CVC4 has known
problems with GCC7 (see issue #1050). Both jobs are added using matrix.include,
based on the example in the documentation:
https://docs.travis-ci.com/user/languages/cpp/#GCC-on-Linux . This
unfortunately requires some code duplication but there does not seem to be a
way to do it in a much better fashion.